Create webscrping
Hi Guys,

I was able to figure out the [basic] script for web-scraping. Please note that the above code works for webpages that only have 1 table.  Since Basketball-reference has webpages with multiple tables, inspection of the html code is required to identify the specific table- I'm working on that now.